pyproject.toml
[project]
name = "em-bot"
version = "0.1.0"
description = "EM Vidros RAG AI Assistant - Production-ready chatbot with memory and vector search"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "inngest>=0.5.15",
    "llama-index-core>=0.14.13",
    "llama-index-readers-file>=0.5.6",
    "llama-index-vector-stores-qdrant>=0.5.0",
    "llama-index-embeddings-openai>=0.3.0",
    "llama-index-llms-openai>=0.3.0",
    "openai>=2.16.0",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "streamlit>=1.53.1",
    "uvicorn>=0.40.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.32.0",
    "aiohttp>=3.11.0",
    "aiosqlite>=0.21.0",
    "pydantic>=2.10.0",
    "python-multipart>=0.0.20",
    "pypdf>=5.0.0",
    "python-docx>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9.0",
    "pytest>=8.3.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]
